Use /etc/spryker-vm-env as a copy of .vm file
We make a copy of /vagrant/.vm during provisioning of machine
at developer's workstation, when he does "vagrant up". This is the moment
where contents of .vm file will be populated. We can't do it in systemd dropin,
as (1) fpm can be started before /vagrant is mounted and (2) because ExecStartPre
is executed after EnvironmentFile is loaded by systemd.